+++ src/color.rs
@@ -1,4 +1,6 @@
-use crossterm::style::Color;
+use {
+    crossterm::style::Color,
+};
 
 /// Build a RGB color
 ///
@@ -10,8 +12,8 @@ pub const fn rgb(r: u8, g: u8, b: u8) -> Color {
 }
 
 /// Build a gray-level color, from 0 (mostly dark) to 23 (light).
-pub fn gray(level: u8) -> Color {
-    assert!(level < 24, "invalid gray level (must be in 0..24)");
+pub fn gray(mut level: u8) -> Color {
+    level = level.min(23);
     Color::AnsiValue(0xE8 + level)
 }
 
@@ -20,16 +22,3 @@ pub const fn ansi(level: u8) -> Color {
     Color::AnsiValue(level)
 }
 
-#[cfg(test)]
-mod color_tests {
-
-    use crate::color::*;
-
-    /// check the color range is correctly checked and a meaningful
-    /// error is raised
-    #[test]
-    #[should_panic(expected = "invalid gray level (must be in 0..24)")]
-    fn check_gray_panic() {
-        let _ = gray(24);
-    }
-}