+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 *If you're reading this because you try make sense of some new API or a breaking change, you might also be interested in coming to the chat for explanations or guidance.*
 
+### next
+- gray panicking on out of range level has been changed to a silent clamping
+
 <a name="v0.23.1"></a>
 ### v0.23.1 - 2023-06-15
 - reexport coolor to ease dependency managment